refactor: move radar controls to header
